pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stimloss"
version = "0.1.1"
description = "Stimloss: power-loss & efficiency analysis for multichannel electrical stimulation."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Francesc Varkevisser and contributors"}]
dependencies = [
  "numpy>=1.26",
  "scipy>=1.11",
  "pandas>=2.1",
  "matplotlib>=3.8",
  "seaborn>=0.13",
  "pyarrow>=16.0.0",
  "typer>=0.12",
  "openpyxl>=3.1",
  "PyYAML>=6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
  "mypy>=1.10",
]

[project.scripts]
stimloss = "stimloss.cli:main"
